Convert image to qcow2
Closes #3

+++ create-image.sh
@@ -5,14 +5,14 @@ set -ex
 # if HDA_SIZE starts with "+", add to computed size.
 if [[ $HDA_SIZE == +* ]]; then
   BASE_SIZE=$(du -s "$CHROOT" | awk "{print int(\$1)}")
-  qemu-img create "$HDA" "${BASE_SIZE}k"
-  qemu-img resize "$HDA" "$HDA_SIZE"
+  qemu-img create -f raw "$HDA_RAW" "${BASE_SIZE}k"
+  qemu-img resize -f raw "$HDA_RAW" "$HDA_SIZE"
 else
-  qemu-img create "$HDA" "$HDA_SIZE"
+  qemu-img create "$HDA_RAW" "$HDA_SIZE"
 fi
 
-mkfs.ext2 -d "$CHROOT/" "$HDA"
+mkfs.ext2 -d "$CHROOT/" "$HDA_RAW"
+[ -w "$CHROOT" ] && rm -rf "$CHROOT"
 
-if [ -w "$CHROOT" ]; then
-  rm -rf "$CHROOT"
-fi
+qemu-img convert -f raw -O qcow2 "$HDA_RAW" "$HDA_QCOW2"
+[ -w "$HDA_RAW" ] && rm "$HDA_RAW"
